refactor(home-auth): refactor out home and auth
