+++ src/components/timer/Controls.js
