+++ src/components/timer/Selections.js
