+++ src/components/timer/TimerInfo.js
